requirements.txt
Django==5.0.14
psycopg[binary]==3.3.3
gunicorn==23.0.0
whitenoise==6.9.0
django-unfold==0.63.0
